+++ aws-cs-webserver/README.md
@@ -49,8 +49,8 @@ To deploy your infrastructure, follow the below steps.
     $ pulumi stack output
     Current stack outputs (2):
         OUTPUT          VALUE
-        PublicDns       ec2-34-217-176-141.us-west-2.compute.amazonaws.com
-        PublicIp        34.217.176.141
+        publicDns       ec2-34-217-176-141.us-west-2.compute.amazonaws.com
+        publicIp        34.217.176.141
     ```
 
 1.  Verify that the EC2 instance exists, by either using the AWS Console or running `aws ec2 describe-instances`.
